Fixed Swapping Creature Bug
cause: loop & scanner.close()

+++ src/Creature.java
@@ -1,8 +1,7 @@
-import java.util.ArrayList;
 
 public class Creature {
     
-    private double nHealth;    
+    private int nHealth;    
     private int nImageID;    
     private String strCreatureName;    
     private char cFamily;    
@@ -10,7 +9,7 @@ public class Creature {
     private int nLevel;    
     private boolean bActive;    
 
-    public Creature(double nHealth, int nImageID, String strCreatureName, char cFamily, String strType, int nLevel, boolean bActive){
+    public Creature(int nHealth, int nImageID, String strCreatureName, char cFamily, String strType, int nLevel, boolean bActive){
         this.nHealth = nHealth;
         this.nImageID = nImageID;
         this.strCreatureName = strCreatureName;
@@ -20,11 +19,11 @@ public class Creature {
         this.bActive = bActive;
     }
 
-    public void setHealth(double nHealth){
+    public void setHealth(int nHealth){
         this.nHealth = nHealth;
     }
     
-    public double getHealth(){
+    public int getHealth(){
         return nHealth;
     }
